requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-allauth==0.54.0
psycopg2==2.9.5
pycodestyle==2.10.0
python-dotenv==1.0.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
pyyaml==6.0
Pillow==9.4.0
coverage==7.2.3
djoser==2.2.0
requests
python-dateutil
beautifulsoup4
django-filter==23.5